improved alert implementation
