+++ packages/ui/src/main.tsx
@@ -6,7 +6,8 @@ import ReactDOM from "react-dom/client";
 import { lightTheme, RainbowKitProvider } from "@rainbow-me/rainbowkit";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { WagmiProvider } from "wagmi";
-import { AlertProvider } from "./contexts/alert-context";
+import ErrorAlert from "./components/error-alert";
+import { AlertProvider } from "./contexts/alert/alert-provider";
 import { Web3ClientProvider } from "./contexts/web3-client-context";
 import config from "./lib/wagmi-config";
 import "./main.css";
@@ -35,6 +36,7 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
         >
           <Web3ClientProvider>
             <AlertProvider>
+              <ErrorAlert />
               <RouterProvider router={router} />
             </AlertProvider>
           </Web3ClientProvider>
