+++ packages/ui/src/routes/withdraw.tsx
@@ -1,5 +1,5 @@
 import EthIcon from "@/assets/ethereum-icon.svg";
-import { useAlertContext } from "@/contexts/alert-context";
+import { useAlertContext } from "@/contexts/alert/alert-context";
 import { useWeb3ClientContext } from "@/contexts/web3-client-context";
 import useArbitrumBridge, { ClaimStatus } from "@/hooks/useArbitrumBridge";
 import {
